refactor: Calculator 클래스  divide 메서드 -> double 캐스팅

+++ src/lv2/Calculator.java
@@ -41,7 +41,7 @@ public class Calculator {
         if (num2 == 0) {
             throw new ArithmeticException("나눗셈 연산에서 분모(두번째 정수)에 0이 입력될 수 없습니다.");
         } else {
-            result = num1 / num2;
+            result = (double) num1 / num2;
         }
         return result;
     }
